test: move fidelity tests to a dedicated Swift Testing file (cmux policy)
cmux test-framework policy: new non-UI tests use Swift Testing, not XCTest.
The fidelity tests were appended to the XCTest file WorkspaceUnitTests.swift;
move them to a dedicated cmuxTests/MobileWorkspaceListFidelityTests.swift
written in Swift Testing (#expect/#require/@Test/@Suite(.serialized)), wired
into the cmux-unit target via project.pbxproj. WorkspaceUnitTests.swift is
restored to base (its XCTest import stays only for its pre-existing suites,
which we must not bulk-rewrite). Same behavioral assertions; the prior two
commits keep the XCTest red/green proof in history.
