+++ cmuxTests/WorkspaceUnitTests.swift
@@ -7212,139 +7212,3 @@ final class ExtensionWorktreePrototypeTests: XCTestCase {
         return output
     }
 }
-
-/// Covers the mobile workspace-list fidelity fixes: terminals are serialized in
-/// the on-screen bonsplit spatial order, a terminal rename re-emits to the phone,
-/// and a pure drag-reorder is detected even though it changes no panel-set state.
-@MainActor
-final class MobileWorkspaceListFidelityTests: XCTestCase {
-    /// Builds a workspace with `count` terminals as tabs in a single pane so that
-    /// a within-pane `reorderTab` genuinely changes their on-screen order. Returns
-    /// the workspace and panel ids in spatial (tab) order.
-    private func makeWorkspaceWithTabTerminals(count: Int) throws -> (Workspace, [UUID]) {
-        precondition(count >= 1)
-        let manager = TabManager()
-        let workspace = try XCTUnwrap(manager.selectedWorkspace)
-        var orderedIds: [UUID] = [try XCTUnwrap(workspace.focusedPanelId)]
-        for _ in 1..<count {
-            let panel = try XCTUnwrap(workspace.newTerminalSurfaceInFocusedPane(focus: false))
-            orderedIds.append(panel.id)
-        }
-        return (workspace, orderedIds)
-    }
-
-    /// Builds a workspace with `count` terminals laid out left-to-right via
-    /// horizontal splits (each in its own pane), returning the workspace and panel
-    /// ids in spatial order.
-    private func makeWorkspaceWithSplitTerminals(count: Int) throws -> (Workspace, [UUID]) {
-        precondition(count >= 1)
-        let manager = TabManager()
-        let workspace = try XCTUnwrap(manager.selectedWorkspace)
-        var orderedIds: [UUID] = [try XCTUnwrap(workspace.focusedPanelId)]
-        for _ in 1..<count {
-            let previous = try XCTUnwrap(orderedIds.last)
-            let panel = try XCTUnwrap(
-                workspace.newTerminalSplit(from: previous, orientation: .horizontal, focus: false)
-            )
-            orderedIds.append(panel.id)
-        }
-        return (workspace, orderedIds)
-    }
-
-    func testOrderedPanelIdsMatchesBonsplitSpatialOrder() throws {
-        let (workspace, createdOrder) = try makeWorkspaceWithSplitTerminals(count: 3)
-
-        // orderedPanelIds is derived from bonsplit's left-to-right tab ordering.
-        let ordered = workspace.orderedPanelIds
-        XCTAssertEqual(Set(ordered), Set(createdOrder), "should contain exactly the created panels")
-
-        // It must equal bonsplit's own allTabIds mapping (the spatial source of truth),
-        // not dictionary/UUID order.
-        let expected = workspace.bonsplitController.allTabIds.compactMap {
-            workspace.panelIdFromSurfaceId($0)
-        }
-        XCTAssertEqual(ordered, expected)
-    }
-
-    func testReorderingTerminalsChangesObserverHashAndBumpsLayoutVersion() throws {
-        // Tabs in one pane so a within-pane reorder genuinely changes their order.
-        let (workspace, ordered) = try makeWorkspaceWithTabTerminals(count: 3)
-        XCTAssertEqual(ordered.count, 3)
-
-        let versionBefore = workspace.paneLayoutVersion
-        // Capture the live spatial order right before the reorder rather than
-        // relying on creation order, so the assertion can't be fooled by insertion
-        // semantics.
-        let beforeOrder = workspace.orderedPanelIds
-        XCTAssertEqual(Set(beforeOrder), Set(ordered))
-        let before = MobileWorkspaceListObserver.summaryHashForTesting(
-            tabs: [workspace],
-            selectedTabID: workspace.id
-        )
-
-        // Move the first terminal (in live order) to the end. Same panel set,
-        // different spatial order.
-        let firstTabId = try XCTUnwrap(workspace.surfaceIdFromPanelId(beforeOrder[0]))
-        XCTAssertTrue(workspace.bonsplitController.reorderTab(firstTabId, toIndex: beforeOrder.count - 1))
-
-        // Sanity: the id set is unchanged, but the order changed.
-        let afterOrder = workspace.orderedPanelIds
-        XCTAssertEqual(Set(afterOrder), Set(beforeOrder))
-        XCTAssertNotEqual(afterOrder, beforeOrder, "reorder should change the ordered sequence")
-
-        // The reorder must wake the observer (bonsplit selection state is not
-        // @Published, so paneLayoutVersion is the only signal).
-        XCTAssertGreaterThan(
-            workspace.paneLayoutVersion,
-            versionBefore,
-            "a pure reorder must bump paneLayoutVersion so the observer re-evaluates"
-        )
-
-        let after = MobileWorkspaceListObserver.summaryHashForTesting(
-            tabs: [workspace],
-            selectedTabID: workspace.id
-        )
-        XCTAssertNotEqual(before, after, "a pure reorder must change the mobile summary hash")
-    }
-
-    func testRenamingTerminalChangesObserverHashAndDisplayedTitle() throws {
-        let (workspace, ordered) = try makeWorkspaceWithTabTerminals(count: 2)
-        let panelId = try XCTUnwrap(ordered.first)
-
-        let before = MobileWorkspaceListObserver.summaryHashForTesting(
-            tabs: [workspace],
-            selectedTabID: workspace.id
-        )
-
-        // A terminal rename sets panelCustomTitles (not panelTitles); the observer
-        // must still detect it, and panelTitle must resolve to the custom title that
-        // the mobile workspace.list response serializes.
-        workspace.setPanelCustomTitle(panelId: panelId, title: "Renamed Terminal")
-        XCTAssertEqual(workspace.panelTitle(panelId: panelId), "Renamed Terminal")
-
-        let after = MobileWorkspaceListObserver.summaryHashForTesting(
-            tabs: [workspace],
-            selectedTabID: workspace.id
-        )
-        XCTAssertNotEqual(before, after, "a terminal rename must change the mobile summary hash")
-    }
-
-    func testRenamingWorkspaceChangesObserverHashAndDisplayedTitle() throws {
-        let (workspace, _) = try makeWorkspaceWithTabTerminals(count: 1)
-
-        let before = MobileWorkspaceListObserver.summaryHashForTesting(
-            tabs: [workspace],
-            selectedTabID: workspace.id
-        )
-
-        workspace.setCustomTitle("Renamed Workspace")
-        // The mobile workspace.list response sends workspace.title.
-        XCTAssertEqual(workspace.title, "Renamed Workspace")
-
-        let after = MobileWorkspaceListObserver.summaryHashForTesting(
-            tabs: [workspace],
-            selectedTabID: workspace.id
-        )
-        XCTAssertNotEqual(before, after, "a workspace rename must change the mobile summary hash")
-    }
-}
